Refactor file check
